Refactor fat controller into multiple controllers
