+++ Gemfile
@@ -2,6 +2,7 @@ source "https://rubygems.org"
 
 gem 'sinatra'
 gem 'sinatra-flash'
+gem 'sinatra-partial'
 gem 'data_mapper', '1.2.0'
 gem 'dm-postgres-adapter'
 gem 'bcrypt'
