+++ app/views/_welcome.erb
@@ -1,0 +1,13 @@
+<% if not defined? @signin_page %>
+  <% if current_user %>
+    <h2>Welcome <%= @current_user.username %></h2>
+    <form action="/sessions" method="post">
+      <input type="hidden" name="_method" value="DELETE">
+      <input type="submit" value="Sign out">
+    </form>
+  <% else %>
+  <form action="/sessions/new">
+    <input type="submit" value="Sign in">
+  </form>
+  <% end %>
+<% end %>
